+++ examples/tests/bessel.hpp
@@ -3,7 +3,7 @@
 // Created Date: 11/05/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 30/05/2022
+// Last Modified: 28/06/2022
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -13,15 +13,14 @@
 
 #include <autd3.hpp>
 
-template <typename T>
-void bessel_test(autd3::ControllerX<T>& autd) {
+void bessel_test(autd3::Controller& autd) {
   autd3::SilencerConfig config;
   autd.send(config);
 
   autd3::modulation::Sine m(150);  // 150Hz AM
 
   const autd3::Vector3 apex = autd.geometry().center();
-  autd3::gain::BesselBeam<T> g(apex, autd3::Vector3::UnitZ(), 13.0 / 180.0 * autd3::pi);
+  autd3::gain::BesselBeam g(apex, autd3::Vector3::UnitZ(), 13.0 / 180.0 * autd3::pi);
 
   autd.send(m, g);
 }